backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
httpx==0.27.0
python-dotenv==1.0.1

# Data
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
fredapi==0.5.2

# AI (Groq — free tier, 14,400 req/day)
groq>=1.1.0

# News
feedparser==6.0.11

# Cache
diskcache==5.6.3
